Add .ci folder to host all customization
Today the CI carries custom Makefiles and overwrites the Makefile in the
repo, making it impossible to test updates to Makefiles themselves. Also
these are currently opaque.

Instead we make note of the diffs each scenario needs and include them
in .mk files for cp and dp as per #18. This way developers can see how
the CI tests. This will also allow to test changes to the base
Makefiles.

In a followo-on commit or PR we can even add config files and even the
scripts CI uses to run the tests. This way we make CI repo minimal, and
number of opaque things zero.

Signed-off-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>