Fix for issue #421:Retransmission Timer validation error
"#" is not supported for comment out parmeters in cfg files.
";" has to use for commenting the parameters in the configuration file.
We are used the dpdk file parser to read a configuration file,
as per the dpdk lib docs for parser they hardcoded character ";" to comment out parameters in the file.