Static address allocation support by PDN GW
SPGW takes command line argument of static pool if there is signle DP.
If multiple DP then command line argument is ignored and static pool
coming through app_config.cfg file is used.
When CSReq comes with non zero address, then address is validated to be
present in the range. If yes then CSReq is accepted and the same address
is returned back to MME in CSRsp.